Moved all parsing related files to new directory parser/, some updates towards enums and parsing
