--- src/game/enums/args/bird_power_args.py
@@ -1,17 +1,0 @@
-from enum import Enum, auto
-
-
-class Entailment(Enum):
-    MAY = auto()
-
-
-class Item(Enum):
-    IT = auto()
-    PREVIOUS = '^'
-
-
-class Location(Enum):
-    BIRDFEEDER = auto()
-    SUPPLY = auto()
-    THIS = auto()
-
