+++ src/parser/enums/__init__.py
@@ -1,0 +1,2 @@
+class ComponentType:
+    pass
