+++ src/parser/enums/base.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class AutoName(Enum):
+    def _generate_next_value_(name, start, count, last_values):
+        return name
